refactor(timetable-notification): separate parse title & body
