+++ src/main/java/com/dongsoop/dongsoop/timetable/notification/TimetableNotification.java
@@ -1,10 +1,8 @@
 package com.dongsoop.dongsoop.timetable.notification;
 
-import com.dongsoop.dongsoop.timetable.dto.TodayTimetable;
 import java.util.List;
-import java.util.Map;
 
 public interface TimetableNotification {
 
-    void send(Long memberId, List<TodayTimetable> timetables, Map<Long, List<String>> deviceByMember);
+    void send(String title, String body, List<String> devices);
 }
