+++ src/main/java/com/dongsoop/dongsoop/timetable/schedule/TimetableSchedule.java
@@ -8,6 +8,7 @@ import com.dongsoop.dongsoop.timetable.repository.TimetableRepository;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.Year;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -18,6 +19,8 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 public class TimetableSchedule {
+
+    private final static String TITLE_FORMAT = "[시간표 알림] 오늘 %d개의 수업이 있습니다";
 
     private final TimetableRepository timetableRepository;
     private final TimetableNotification timetableNotification;
@@ -41,7 +44,14 @@ public class TimetableSchedule {
         Map<Long, List<String>> deviceByMember = memberDeviceService.getDeviceByMember(targetsIdList);
 
         timetableByMember.forEach(
-                (memberId, timetables) -> timetableNotification.send(memberId, timetables, deviceByMember));
+                (memberId, timetables) -> {
+                    String title = String.format(TITLE_FORMAT, timetables.size());
+                    String body = getTimetableBody(timetables);
+
+                    List<String> devices = deviceByMember.getOrDefault(memberId, List.of());
+
+                    timetableNotification.send(title, body, devices);
+                });
     }
 
     /**
@@ -58,5 +68,26 @@ public class TimetableSchedule {
         SemesterType semester = SemesterType.fromMonth(month);
 
         return timetableRepository.getTimetableNotificationDtoList(year, semester, dayOfWeek);
+    }
+
+    /**
+     * 시간표 본문 생성
+     *
+     * @param timetables 시간표 리스트
+     * @return 시간표 본문
+     */
+    private String getTimetableBody(List<TodayTimetable> timetables) {
+        StringBuilder stringBuilder = new StringBuilder();
+        timetables.forEach(timetable -> {
+            String timeFormat = timetable.startAt()
+                    .format(DateTimeFormatter.ofPattern("HH:mm"));
+            stringBuilder.append(timeFormat)
+                    .append(" ")
+                    .append(timetable.name())
+                    .append("\n");
+        });
+
+        return stringBuilder.toString()
+                .trim();
     }
 }
